image_analysis/app/requirements.txt
google-cloud-vision==0.40.0
opencv-python==4.1.2.30
matplotlib==3.1.2
imagesize==1.1.0
pandas==0.25.3
pillow
flask
flask_cors
